fix: baseline budget
